requirements.txt
# Requirements for Job Scheduler Takii Web Application

Flask==3.0.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flasgger==0.9.7.1